+++ include/ccf/tx_status.h
@@ -31,6 +31,13 @@ namespace ccf
     Invalid,
   };
 
+  // Contains only the terminal values of TxStatus
+  enum class FinalTxStatus : std::underlying_type_t<TxStatus>
+  {
+    Committed = static_cast<std::underlying_type_t<TxStatus>>(TxStatus::Committed),
+    Invalid = static_cast<std::underlying_type_t<TxStatus>>(TxStatus::Invalid),
+  };
+
   constexpr char const* tx_status_to_str(TxStatus status)
   {
     switch (status)